Made Attack functional and all skeletons move
Made it so when player is attacking and the hitbox collides with the skeleton hitbox, the skeleton gets killed. I finalised the map design and made it so spikes surround the border. Also I added a loop that adds a steplistener for every skeleton so now all the skeletons can move.