Finalised my start screen + fixed a bug with crouching
Made it so that buttons have images attached to them, clicking the Play and Quit buttons makes a sound and while in the menu, menu music plays. Also made it so that while moving you can't crouch as before, you could crouch and for some reason that made it so the player was sliding a lot too which made it really fast.